requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.4
django-countries==7.4.2
django-import-export==3.0.1
django-user-agents==0.4.0
docopt==0.6.2
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
numpy==1.23.5
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.2
pipreqs==0.4.11
pygeoip==0.3.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
razorpay==1.3.0
requests==2.28.1
scrapetube==2.3.1
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
typing_extensions==4.4.0
ua-parser==0.16.1
urllib3==1.26.13
user-agents==2.2.0
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
django-hosts